+++ gmlx/load/modules.py
@@ -1025,7 +1025,7 @@ def _make_fused_block(base_cls, caps):
                 gates = mx.softmax(
                     self.gate(xf), axis=-1, precise=True)
                 k = self.top_k
-                inds = mx.argpartition(gates, kth=-k, axis=-1)[..., -k:]
+                inds = mx.stop_gradient(mx.argpartition(gates, kth=-k, axis=-1)[..., -k:])
                 scores = mx.take_along_axis(gates, inds, axis=-1)
                 if self.norm_topk_prob:
                     scores = scores / scores.sum(axis=-1, keepdims=True)