+++ gmlx/models/qwen4_exp/model.py
@@ -1348,7 +1348,7 @@ class SparseMoeBlock(nn.Module):
         gates = self.gate(x)
         gates = mx.softmax(gates, axis=-1, precise=True)
         k = self.top_k
-        inds = mx.argpartition(gates, kth=-k, axis=-1)[..., -k:]
+        inds = mx.stop_gradient(mx.argpartition(gates, kth=-k, axis=-1)[..., -k:])
         scores = mx.take_along_axis(gates, inds, axis=-1)
         if self.norm_topk_prob:
             scores = scores / scores.sum(axis=-1, keepdims=True)